### 2019-02-08 #1241 @whedon
[PRE REVIEW]: Optimising Light Source Positioning for Even and Flux-Efficient Illumination 
**Submitting author:** @adrena-lab (<a href="http://orcid.org/0000-0002-4413-5527">Stefan  Guldin</a>)
**Repository:** <a href="https://github.com/adrena-lab/Optimising-Light-Source-Positioning" target ="_blank">https://github.com/adrena-lab/Optimising-Light-Source-Positioning</a>
**Version:** 1.0
