### 2019-02-08 #1241 @arfon
@adrena-lab - please rename your paper to `paper.md`. In addition, I don't see a bibtex file in this repository.
