+++ Bibliography.bib
@@ -1,0 +1,19 @@
+@incollection{Pedrotti2018,
+address = {Cambridge},
+author = {Pedrotti, Frank L and Pedrotti, Leno M and Pedrotti, Leno S},
+booktitle = {Introduction to Optics},
+chapter = {1},
+edition = {3},
+pages = {1--15},
+publisher = {Cambridge University Press},
+title = {{Nature of Light}},
+year = {2018}
+}
+
+
+@misc{Convhull,
+keywords = {Convhull},
+mendeley-tags = {Convhull},
+title = {{Convhull - MATLAB {\&} Simulink - MathWorks, United Kingdom}},
+howpublished = {https://uk.mathworks.com/help/matlab/ref/convhull.html (accessed Feb 25, 2019)}
+}
