+++ paper.md
@@ -28,14 +28,14 @@ Furthermore, we demonstrate an approach by which a user can reduce the amount of
 
 Our model describes positions of a light source in Cartesian coordinates (x, y and z) relative to the centre of the illuminated area. The spatial region available for light source placement must be defined by the user. In addition, the user defines a range of allowed angles of illumination in Polar coordinates (theta and phi) relative to the surface normal (see Figure 1a).
 Based on these user inputs, the software creates a set of illumination configurations, which are then tested for feasibility as defined by the user e.g. to filter-out configurations that would result in the light source being positioned in the target-to-camera light path. This pre-filtering process can also be augmented by the user to suit the needs of their particular system.
-The illumination profile as observed by the camera is calculated for each of the positional configurations defined above.  In order to perform this calculation, the model requires the user to specify a far field profile for their specific light source (we provide the characteristics of a 5 mm diameter hemispherical LED in our example code). Our code accounts for the drop in relative intensity due to distance from the light source and angle from the principal axis of the light source. It then utilises a cosine flux correction to identify the projected area difference between the light path and surface normal. Following this, Lambert’s cosine emission law (and a secondary target-to-lens distance factor) is applied in order to determine the radiant intensity as observed by the imaging objective. The code exploits positional symmetry so that illumination from multiple light sources can be computed efficiently.
+The illumination profile as observed by the camera is calculated for each of the positional configurations defined above.  In order to perform this calculation, the model requires the user to specify a far field profile for their specific light source (we provide the characteristics of a 5 mm diameter hemispherical LED in our example code). Our code accounts for the drop in relative intensity due to distance from the light source and angle from the principal axis of the light source. It then utilises a cosine flux correction to identify the projected area difference between the light path and surface normal. Following this, Lambert’s cosine emission law (and a secondary target-to-lens distance factor) is applied in order to determine the radiant intensity as observed by the imaging objective [@Pedrotti2018]. The code exploits positional symmetry so that illumination from multiple light sources can be computed efficiently.
 
 ![](Figure1.png)
 **Figure 1:**  Light source position and illumination profile. **a)** Schematic illustrating the placement of the light source relative to the illuminated area with all relevant variables. **b)** Illumination profile for a single light source denoted by the circle on the bottom left of the plot. **c)** Illumination profile for a system with four active light sources denoted by the black circles on the plot.  The details of the variable values for these light sources are summarized in a legend on the bottom left of the plot. 
 
 Solving these equations across a matrix of points representing the imaging plane allows one to predict the general illumination pattern and calculate numerical metrics that can be used to optimise the system i.e. total radiant flux (efficiency of light delivered to the target area) and the standard deviation of the flux as a percentage of the mean flux (illumination uniformity). A user can filter the positional configurations dataset using these defined metrics or their own in order to select an optimised illumination configuration for their system. 
 
-We provide a method for filtering these configurations to select settings with the lowest standard deviation for a given total radiant flux.  This method uses a convex hull to determine the set of points required to enclose all other points in the dataset.  As this operation seeks to fully enclose the dataset, a subset of these points will represent regions of low radiant flux which are then removed.  The remaining points represent the best trade-off between total radiant flux and standard deviation (i.e. for a given flux what is the lowest standard deviation possible from the positional configurations dataset).  As standard deviation is not necessarily a perfect descriptor for uniformity, additional data points may be of interest.  Therefore, the user can provide tolerance values for both total radiant flux and standard deviation to broaden the domain of selected configurations.  This process is detailed in Figure 2.  
[…]
