+++ paper.md
@@ -35,7 +35,7 @@ The illumination profile as observed by the camera is calculated for each of the
 
 Solving these equations across a matrix of points representing the imaging plane allows one to predict the general illumination pattern and calculate numerical metrics that can be used to optimise the system i.e. total radiant flux (efficiency of light delivered to the target area) and the standard deviation of the flux as a percentage of the mean flux (illumination uniformity). A user can filter the positional configurations dataset using these defined metrics or their own in order to select an optimised illumination configuration for their system. 
 
-We provide a method for filtering these configurations to select settings with the lowest standard deviation for a given total radiant flux.  This method uses a convex hull to determine the set of points required to enclose all other points in the dataset[@Convhull].  As this operation seeks to fully enclose the dataset, a subset of these points will represent regions of low radiant flux which are then removed.  The remaining points represent the best trade-off between total radiant flux and standard deviation (i.e. for a given flux what is the lowest standard deviation possible from the positional configurations dataset).  As standard deviation is not necessarily a perfect descriptor for uniformity, additional data points may be of interest.  Therefore, the user can provide tolerance values for both total radiant flux and standard deviation to broaden the domain of selected configurations.  This process is detailed in Figure 2.  
+We provide a method for filtering these configurations to select settings with the lowest standard deviation for a given total radiant flux.  This method uses a convex hull to determine the set of points required to enclose all other points in the dataset [@Convhull].  As this operation seeks to fully enclose the dataset, a subset of these points will represent regions of low radiant flux which are then removed.  The remaining points represent the best trade-off between total radiant flux and standard deviation (i.e. for a given flux what is the lowest standard deviation possible from the positional configurations dataset).  As standard deviation is not necessarily a perfect descriptor for uniformity, additional data points may be of interest.  Therefore, the user can provide tolerance values for both total radiant flux and standard deviation to broaden the domain of selected configurations.  This process is detailed in Figure 2.  
 
 ![](Figure2-2.png)
 **Figure 2:**  Demonstration of convex hull based data filtering.  **a)** The figures of merit, total flux  and illumination variation (standard deviation), define the axes against which every parameter-combination is mapped (forming a cloud). The modified convex hull (boundary) of the candidate population is shown in blue.  **b)** A total flux tolerance of 0.05 and a standard deviation tolerance of 10<sup>-6</sup> was applied and the selected data points are shown in blue. 
