### 2019-02-25 #1241 @adrena-lab
Dear Arfon and Daniel,

My apologies, there was a misunderstanding on our end.  We have renamed out paper to paper.md and have added paper.bib to the repository.  We will follow this thread to ensure there are no further delays.

Thank you,

Niamh
