### 2019-02-25 #1241 @danielskatz
As you probably can tell, there are some other files missing, such as a .bib file and a figure.  Feel free to add files, and when you have done so, enter a comment here that is ```@whedon generate pdf``` to try the PDF compilation process again.  Let me know when you are done or stuck.
