### 2019-02-25 #1241 @danielskatz
It would be good to explain the research application of this software - what kinds of researchers will use this software, and for what types of research?
