### 2019-02-25 #1241 @danielskatz
Similarly, the README will need some work - it should talk about the research application(s) of the software, plus a bit more.  See https://joss.readthedocs.io/en/latest/review_criteria.html under "Documentation" and "A statement of need".
