+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Optimising Light Source Positioning for Even and Intense Illumination'
+title: 'Optimising Light Source Positioning for Even and Flux-Efficient Illumination'
 tags:
   - Even Illumination
   - Modelling Illumination
@@ -20,28 +20,26 @@ date: 26 September 2018
 bibliography: paper.bib
 ---
 # Summary
-The position of light sources within an imaging system has significant consequences for the downstream usability of the data it generates. Non-uniform illumination can contribute to low quality (or in some cases unusable) images, particularly so when illumination variation approaches or exceeds the sensitivity range of the capture device/camera. Similarly, low flux efficiency (i.e. flux through imaging plane divided by total flux from source) will negatively affect the image acquisition and subsequent analysis. Often flux efficiency is sacrificed for illumination uniformity i.e. choosing to deliver less light to the target such that the variation of light intensity across the target remains low. The large number of possible positional configurations of a light source within an imaging system precludes manual optimisation. 
+The position of light sources within an imaging system has significant consequences for the downstream usability of the data it generates. Non-uniform illumination can contribute to low quality (or in some cases unusable) images, particularly so when illumination variation approaches or exceeds the sensitivity range of the capture device/camera. Similarly, low flux efficiency (i.e. ratio of flux through the imaging plane divided by the total flux from the source) will negatively affect the image acquisition and subsequent analysis. Often flux efficiency is sacrificed for illumination uniformity i.e. choosing to deliver less light to the target in order to keep the variation of light intensity across the target remains low. Furthermore, the large number of possible positional configurations of a light source within an imaging system precludes manual optimisation. 
 
-To tackle this issue, we offer a software for modelling the illumination profile for a given light source. The code can be easily adjusted to model a variety of positional configuration and can rapidly calculate the results for many thousands of variable combinations.
-
-Furthermore, we demonstrate an approach by which a user can reduce the amount of possible multi-variable combinations down to the most viable options. This is performed using a modified convex hull approach in two-dimensions (optimising for two figures of merit, i.e. the total flux and the illumination variation). In principle, this method can be extended to n-dimensional space to include additional figures of merit for optimisation.
-
-Our model describes positions of a light source in Cartesian coordinates (x, y and z) relative to the centre of the illuminated area. The spatial region available for light source placement must be defined by the user. In addition, the user defines a range of allowed angles of illumination in Polar coordinates (theta and phi) relative to the surface normal (see Figure 1a).
-Based on these user inputs, the software creates a set of illumination configurations, which are then tested for feasibility as defined by the user e.g. to filter-out configurations that would result in the light source being positioned in the target-to-camera light path. This pre-filtering process can also be augmented by the user to suit the needs of their particular system.
-The illumination profile as observed by the camera is calculated for each of the positional configurations defined above.  In order to perform this calculation, the model requires the user to specify a far field profile for their specific light source (we provide the characteristics of a 5 mm diameter hemispherical LED in our example code). Our code accounts for the drop in relative intensity due to distance from the light source and angle from the principal axis of the light source. It then utilises a cosine flux correction to identify the projected area difference between the light path and surface normal. Following this, Lambert’s cosine emission law (and a secondary target-to-lens distance factor) is applied in order to determine the radiant intensity as observed by the imaging objective [@Pedrotti2018]. The code exploits positional symmetry so that illumination from multiple light sources can be computed efficiently.
+To tackle this issue, we offer a software for modelling the illumination profile for a given light source. The code can be easily adjusted to model a variety of positional configuration and rapidly calculates results for many thousands of variable combinations.
 
 ![](Figure1.png)
[…]
