+++ paper.md
@@ -29,7 +29,7 @@ To tackle this issue, we offer a software for modelling the illumination profile
 Furthermore, we demonstrate an approach by which a user can reduce the amount of possible multi-variable combinations down to the most viable options. This is performed using a modified convex hull approach in two-dimensions (optimising for two figures of merit, i.e. the total flux and the illumination variation). In principle, this method can be extended to n-dimensional space to include additional figures of merit for optimisation.
 
 Our model describes positions of a light source in Cartesian coordinates (x, y and z) relative to the centre of the illuminated area. The spatial region available for light source placement must be defined by the user. In addition, the user defines a range of allowed angles of illumination in Polar coordinates (theta and phi) relative to the surface normal (see Figure 1a).
-Based on these user inputs, the software creates a set of illumination configurations.  These are then tested for feasibility based on the requirements and limitations of the imaging device. This pre-filtering process can also be augmented by the user to suit the needs of their particular system.
+Based on these user inputs, the software creates a set of illumination configurations.  These are then tested for feasibility with regard to the requirements and limitations of the imaging device. This pre-filtering process can also be augmented by the user to suit the needs of their particular system.
 The illumination profile as observed by the camera is calculated for each of the positional configurations defined above.  In order to perform this calculation, the model requires the user to specify a far field profile for their specific light source (we provide the characteristics of a 5 mm diameter hemispherical LED in our example code). Our code accounts for the drop in relative intensity due to distance from the light source and angle from the principal axis of the light source. It then utilises a cosine flux correction to identify the projected area difference between the light path and surface normal. Following this, Lambert’s cosine emission law (and a secondary target-to-lens distance factor) is applied in order to determine the radiant intensity as observed by the imaging objective [@Pedrotti2018]. The code exploits positional symmetry so that illumination from multiple light sources can be computed efficiently.
 
 Solving these equations across a matrix of points representing the imaging plane allows one to predict the general illumination pattern and calculate numerical metrics that can be used to optimise the system i.e. total radiant flux (efficiency of light delivered to the target area) and the standard deviation of the flux as a percentage of the mean flux (illumination uniformity). A user can filter the positional configurations dataset using these defined metrics or their own in order to select an optimised illumination configuration for their system. 
