### 2019-02-27 c6b10e5
Delete Convhull_Option_Reduction.m
