+++ Convhull_Option_Reduction.m
@@ -1,182 +1,0 @@
-function [Illumination_Data_Reduced] = Convhull_Option_Reduction(Illumination_Data, Std_Dev_Selector, Total_Flux_Tolerance, Standard_Deviation_Tolerance)
-
-%This function uses a modified convex hull method to select those points which
-%maximize total flux for a given standard deviation. 
-
-%These lines of code implement the user's selection between the absolute
-%standard deviation and the standard deviation as a percentage of
-%illumination. 
-if Std_Dev_Selector == 1
-    std_dev = Illumination_Data.standard_deviation;
-else
-    std_dev = Illumination_Data.standard_deviation_percentage;    
-end
-
-%Defines the Total_Intensity
-I = Illumination_Data.total_flux;
-
-%Loads additional variables
-data_theta = Illumination_Data.theta;
-data_phi = Illumination_Data.phi;
-data_x_position = Illumination_Data.x_position;
-data_y_position = Illumination_Data.y_position;
-data_height = Illumination_Data.height;
-
-data_Total_Flux = Illumination_Data.total_flux;
-data_Std_Dev = Illumination_Data.standard_deviation;
-data_Std_Perc = Illumination_Data.standard_deviation_percentage;
-
-%Sort the data according to standard deviation from smallest to largest
-[std_dev_sort , In] = sort(std_dev);
-I_sort = I(In);
-
-I = I_sort;
-std_dev = std_dev_sort;
-
-%Perform same sort on other variables
-data_theta = data_theta(In);
-data_phi = data_phi(In);
-data_x_position = data_x_position(In);
-data_y_position = data_y_position(In);
-data_height = data_height(In);
-
-data_Total_Flux = data_Total_Flux(In);
-data_Std_Dev = data_Std_Dev(In);
-data_Std_Perc = data_Std_Perc(In);
-
-%These are the tolerances that the user has already defined
-dx = Standard_Deviation_Tolerance;
-dy = Total_Flux_Tolerance;
-
-%Matlab script takes the standard deviation and total intensity as the x
-%and y values respectively.  It then finds the convex hull of the data set,
-%i.e. a subset of points that enclose all other points in the dataset.  
-%K is a vector of indices for these points
-K = convhull(std_dev, I);
-
-%This convex hull function moves in a counter-clockwise manner, as such, it
-%will first select those points which have a low total flux.  We reverse
-%the order of this vector for ease of later calculations. 
-K = flipud(K);
-
-%Uses the indices to extract the relevant points from the standard
-%deviation and intensity data.  
-Ks = std_dev(K);
-KI = I(K);
-
-%The goal is to achieve a high total intensity with a low standard
-%deviation. Therefore, any points that correspond to a high standard deviation
-%and low total intensity can be disregarded.  As the convhull() function
-%chooses points to encompass the entire range of the data, there will be
-%some points that represent a high standard deviation and low intensity.
-%The following lines of code identify these and removes them from the data
-%set.  
-[~, idx_upper] = max(KI);
-
-Ks = Ks(1:idx_upper);
-KI = KI(1:idx_upper);
-K = K(1:idx_upper);
-
-%The indices for all chosen points will be recorded in this vector 
-Index_all = [];
-
-for i = 1:length(Ks)-1
-    
-    %This creates a subset of data that lies between the ith and ith+1
-    %points selected by the convex hull method
-    std_dev_subset = std_dev(K(i):K(i+1)-1,1);
-    intensity_subset = I(K(i):K(i+1)-1,1);
-    
-    %a linear fit is applied to the two convex hull points 
-    p = polyfit([Ks(i), Ks(i+1)],[KI(i), KI(i+1)],1);
-    
[…]
