+++ paper.md
@@ -39,7 +39,7 @@ Solving these equations across a matrix of points representing the imaging plane
 
 We provide a method for filtering these configurations to select settings with the lowest standard deviation for a given total radiant flux.  A convex hull determines the set of points required to enclose all other points in the dataset [@Convhull].  As this operation seeks to fully enclose the dataset, a subset of these points will represent regions of low radiant flux, which are then removed.  The remaining points represent the best trade-off between total radiant flux and standard deviation (i.e. for a given flux, the positional configuration that achieves the lowest possible standard deviation).  As standard deviation is not necessarily a perfect descriptor for uniformity, additional data points may be of interest.  Therefore, the user can provide tolerance values for both total radiant flux and standard deviation to broaden the domain of selected configurations.  This process is detailed in Figure 2.  
 
-![](Figure2.png)
+![](Figure2-2.png)
 **Figure 2:**  Demonstration of convex hull based data filtering.  **a)** The figures of merit, total flux  and illumination variation (standard deviation), define the axes against which every parameter-combination is mapped (forming a cloud). The modified convex hull (boundary) of the candidate population is shown in blue.  **b)** A zoomed-in view of a.  **c)** Tolerance values of 0.05 for total flux and 10<sup>-6</sup> were applied; this allowed for additional points that lay within these tolerances to be selected by the modified convex hull.  
   
 Finally, our supplied plotting function enables the user to visually check the illumination profile and the relative positions of the light sources to ensure the results are acceptable (see Figure 1b&c).  In summary, this software can screen large numbers of possible variable values for the positioning of the light sources with respect to total flux and uniformity. This will allow a user to design an optimised light array for their imaging system without requiring numerous, time-consuming experimental measurements.
