### 2019-02-27 #1241 @adrena-lab
Dear Daniel,

I have adjusted the paper to include the research applications and target audience and provided significant modifications to the ReadMe file.  I hope this is sufficient.

Thank you,

Niamh
