### 2019-03-08 #1241 @trallard
Hi @adrena-lab I will be the editor do you have any reviewers in mind? You can also check the list fo reviewers [here](https://bit.ly/joss-reviewers)
