### 2019-04-16 #1241 @trallard
hi @adrena-lab  sorry for the long wait, I am now back and in full swing so let's get things moving

Dear potential reviewers (@aquilesC @cmccomb @nicoguaro @martlj @PhilipVinc),

It seems that you all have valuable expertise to review this submission, so I was wondering if you had the time and bandwidth to take upon as reviewers?

**Title**: open_iA: Optimising light source positioning for even and flux-efficient illumination

**Summary**: Visual analysis and processing of volumetric datasets, with a focus on industrial computed tomography

Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01139/joss.01139/10.21105.joss.01139.pdf

Submitting author: @adrena-lab

Repository: [https://github.com/adrena-lab/Optimising-Light-Source-Positioning](https://github.com/adrena-lab/Optimising-Light-Source-Positioning)

Version: 1.0
