### 2019-04-16 #1241 @nicoguaro
@trallard, I think I can review it but it will take me a little bit longer than usual.

@adrena-lab, the software seems to be written in MATLAB. Is it mandatory, or can it run in Octave? If it cannot run in Octave, which version of MATLAB is needed?
