### 2019-04-17 #1241 @adrena-lab
Dear All,

I believe the link to the article proof above is incorrect (from a different contribution, not authored by us). Should be this one:

https://github.com/openjournals/joss-papers/blob/joss.01241/joss.01241/10.21105.joss.01241.pdf

Thanks,
Stefan
