### 2019-04-17 #1241 @adrena-lab
@nicoguaro The software was written and tested using Matlab 2018a. It may indeed run using Octave as the code does not rely upon Matlab proprietary toolboxes, however this hasn't been tested by us.
