### 2019-04-18 #1392 @trallard
@nicoguaro @cmccomb thanks for agreeing to review this submission.

You will find a checklist for each of you, please tick the items as you go over them. Use this issue to keep the updates and discussions in a central place. However, if there are longer issues or improvements that need to be requested, please make an issue on the software repo and reference it here 😄

If you encounter any issues or have any questions about the review process feel free to ping me here
