### 2019-04-18 #1392 @cmccomb
@adrena-lab This is a great repository! Here's what I think should be updated/revised:

1. General Checks, Version: You should make a release with version 1.0.
2. Documentation, Installation instructions: Please state the dependencies. For MATLAB, this probably just means to indicate which versions it should run on. If you want to make claims for Octave (recommended!) please provide version information for that too. It would be helpful to provide a short description fo what each file does, as well as a description of what the .mat file is used for.
3. Documentation, Automated Tests: There is not way to check the functionality of the software. Maybe the authors could provide an expected optimum value for the example that users could compare against?
4. Documentation, Example Usage: Will the authors please provide some more description of the example case in the README file? This would be helpful for the reader. The content can be borrowed from the software paper.
5. Documentation, Community Guidelines: Community guidelines should be added. In addition, the zip file should not be used in the repository in order to make it easier for others to contribute. Include the .m and .mat files directly.
