### 2019-04-18 #1392 @nicoguaro
@adrena-lab, I downloaded the repository and ran the main file using Octave 5.1 under Windows 10. I agree with the comments added by @cmccomb.

I will test the software in MATLAB later, in the mean time I add the errors I obtained using Octave.

The first error was
```
error: 'fullfact' undefined near line 44 column 16
error: called from
    Create_Variable_Combinations at line 44 column 14
    Optimising_Light_Source_Positioning at line 62 column 16
```

This was fixed doing the following in the command prompt.

    pkg load statistics

Similarly, I obtained the following error

```
error: 'std2' undefined near line 171 column 28
error: called from
    Illumination_Calculations at line 171 column 26
    Optimising_Light_Source_Positioning at line 158 column 21
```

That was fixed with

    pkg load image

There is also the following error:

```
error: 'newline' undefined near line 176 column 13
error: called from
    Convhull_Option_Reduction at line 176 column 5
    Optimising_Light_Source_Positioning at line 189 column 29
```
