+++ paper.bib
@@ -3,6 +3,7 @@ address = {Cambridge},
 author = {Pedrotti, Frank L and Pedrotti, Leno M and Pedrotti, Leno S},
 booktitle = {Introduction to Optics},
 chapter = {1},
+doi = {10.1017/9781108552493},
 edition = {3},
 pages = {1--15},
 publisher = {Cambridge University Press},
