### 2019-05-02 08cc533
Delete Optimising_Light_Source_Positioning.m
