+++ Optimising_Light_Source_Positioning.m
@@ -1,216 +1,0 @@
-%This is the main file for the optimisation of light source positioning, it
-%will cover each step of the process along with associated instructions on
-%each step and the modifications the user needs to perform in order to
-%tailor the system to their own set-up 
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-%%%    STEP 1    %%% 
-
-%There are five variables for this system.  The position of the light
-%source is defined in Cartesian coordinates relative to the centre of the
-%illuminated area.  Further to this, a range of allowed angles of 
-%illumination in Polar coordinates (theta and phi) relative to the surface 
-%normal must be provided by the user. Please see Figure 1a for a schematic
-%demonstrating these variables.  
-
-%NOTE: These values are for one light source, in the
-%Illumination_Calculations step (Step 3) it is assumed that there are four
-%identical light sources which are placed around the illuminated area over 
-%two lines of symmetry. It is noted in Step 3 where the user may adjust this.  
-
-
-%The user must adjust these values accordingly:
-
-%Theta, enter in degrees
-theta_range_interval = 45;
-theta_range_lower_limit = 0;
-theta_range_upper_limit = 85;
-
-%Phi, enter in degrees
-phi_range_interval = 45;
-phi_range_lower_limit = 0;
-phi_range_upper_limit = 90;
-
-%X-Position
-x_interval = 45;
-x_lower_limit = -100;
-x_upper_limit = -10;
-
-%Y-Position 
-y_interval = 35;
-y_lower_limit = -80;
-y_upper_limit = -10;
-
-%Height above illuminated surface (z-position)
-H_interval = 50;
-H_lower_limit = 10;
-H_upper_limit = 110;
-
-%The values provided by the user are collected here into a matrix of input
-%variables which will be used to determine all possible combinations
-Input_Variables = [theta_range_interval, theta_range_lower_limit, theta_range_upper_limit;...
-    phi_range_interval, phi_range_lower_limit, phi_range_upper_limit;...
-    x_interval, x_lower_limit, x_upper_limit;...
-    y_interval, y_lower_limit, y_upper_limit;...
-    H_interval, H_lower_limit, H_upper_limit];
-
-%This function file takes the Input_Variables matrix and outputs a matrix
-%of all possible combinations.  Each row represents a possible combination
-%while each column represents one of the five variables, theta, phi,
-%x-position, y-position, and height in that order.
-[Combinations] = Create_Variable_Combinations(Input_Variables);
-
-disp('Step 1 is Complete')
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-%%%    STEP 2    %%% 
-
-%Depending upon the lighting system, some of the combinations will be 
-%unfeasible.  This may be due to design limitations such as providing an 
-%unobstructed view for an imaging device or because certain combinations 
-%are inherently inefficient and removing them will reduce the computing time.  
-
-%As this process is highly individual to each lighting system, we
-%have provided some general protocols for the removal of unsuitable
-%combinations.  These filters can be augmented by the user and additional 
-%screening processes can be added as needed.  It should be noted that this 
-%step is not integral to the overall code and can be neglected.  
-
-%NOTE: Do NOT use this section without reviewing the file and adjusting the
-%values as needed. 
-
[…]
