### 2019-05-02 85b9ed3
Delete Create_Variable_Combinations.m
