### 2019-05-02 64bd5ae
Delete Exclude_Unallowed_Combinations.m
