+++ Plot_Save_Results.m
@@ -1,184 +1,0 @@
-function Plot_Save_Results(Illumination_Data, Plotted_Area_Limits, Illuminated_Area_Limits, Far_Field_Data, Height_limit)
-    
-X_limit = Plotted_Area_Limits(1);
-X_spacing = Plotted_Area_Limits(2);
-Y_limit = Plotted_Area_Limits(3);
-Y_spacing = Plotted_Area_Limits(4);
-
-%Formation of the grid
-x_grid = -1 * X_limit:X_spacing:X_limit;
-y_grid = -1 * Y_limit:Y_spacing:Y_limit;
-
-    X = zeros(1, length(x_grid)-1);
-    Y = zeros(1, length(y_grid)-1);
-
-    for i = 1:length(x_grid)-1
-        for j = 1:length(y_grid)-1
-
-            x = (x_grid(i) + x_grid(i+1))/2;
-            y = (y_grid(j) + y_grid(j+1))/2;
-            X(i) = x;
-            Y(j) = y;
-        end
-    end
-    
-    
-%The far field data is read in here
-Far_Field_x = Far_Field_Data(:,1);
-Far_Field_y = Far_Field_Data(:,2);
-
-S = size(Illumination_Data.theta);
-number_combinations = S(1);
-
-for option = 1:number_combinations
-    
-    %Selects correct value for each variable for each combination
-    theta = Illumination_Data.theta(option);
-    phi = Illumination_Data.phi(option);
-    x_position = Illumination_Data.x_position(option);
-    y_position = Illumination_Data.y_position(option);
-    height = Illumination_Data.height(option);
-    std_data = Illumination_Data.standard_deviation(option);
-    std_perc = Illumination_Data.standard_deviation_percentage(option);
-    flux_data = Illumination_Data.total_flux(option);
-
-
-    %Determines the position the centre of the light beam strikes the
-    %illuminated area
-    d_along_plate = tan(theta)*height;
-    y_along_plate = sin(phi)*d_along_plate;
-    x_along_plate = cos(phi)*d_along_plate;
-
-    light_centre_x = x_position + x_along_plate;
-    light_centre_y = y_position + y_along_plate;
-
-    %Pre-allocates the flux grid 
-    Flux = zeros(length(y_grid)-1, length(x_grid)-1);
-
-    for i = 1:length(x_grid)-1
-        for j = 1:length(y_grid)-1
-
-            x = (x_grid(i) + x_grid(i+1))/2;
-            y = (y_grid(j) + y_grid(j+1))/2;
-
-            %Determines the distance from the source
-            true_distance_from_source = pdist2([x, y, 0], [x_position, y_position, height]);
-            
-            %Determines the reduction in intensity from the source due to
-            %distance
-            Intensity_distance = 1/(true_distance_from_source)^2;
-
-            %Determines the angle from the central beam of light to the
-            %grid position
-            v1 = [x, y, 0] - [x_position, y_position, height];
-            v2 = [light_centre_x, light_centre_y, 0] - [x_position, y_position, height];
-            CosTheta = dot(v1,v2)/(norm(v1)*norm(v2));
-            Angle_from_source = acos(CosTheta);
-
-            %Converts the angle to degrees 
-            Angle_from_source_deg = Angle_from_source*180/pi;
-            
-            %Locates the points nearest to this angle in the far field
-            %data and linearly interpolates as needed. If the angle exceeds that provided by the far field
-            %data, the value of the relative intensity drop is assumed to
[…]
