### 2019-05-02 ab515e8
Delete Optimising_Light_Source_Positioning.m
