+++ README.md
@@ -140,5 +140,8 @@ Standard_Deviation_Tolerance = 10^8;
 This process further reduces the number of viable configurations from 260 to 27 as shown in Figure 2. 
 
 ![](https://github.com/adrena-lab/Optimising-Light-Source-Positioning/blob/Code/Figures/Convhull_reduction.png)
-**Figure 2**  The figures of merit,
+**Figure 2:**  The figures of merit,
 total flux and illumination variation (standard deviation), define the axes against which every parameter-combination is mapped (forming a cloud). The modified convex hull (boundary) of the candidate population is shown in blue and circled.  The values which have been selected based on the user provided tolerances are shown in blue.  
+
+**Step 5**
+The remaining options provided after the modified convhull method was applied can be further reduced based upon user restrictions for minimum total flux or maximum standard deviation.  Furthermore, these final configurations are few enough in number to be plotted without requiring undue computational resources and evaluated by the user.  
