### 2019-05-03 e77d5dc
Update README.md
Updated README.md as per reviewer comments:

Added:
List of dependencies
Community Guidelines
Example Usage
Automated Test
