+++ README.md
@@ -22,7 +22,7 @@ This is the main function file for the software.  In this file, the user provide
 
 **Create_Variable_Combinations.m**
 
-The position of the light source is defined in Cartesian coordinates relative to the centre of the illuminated area.  Further to this, a range of allowed angles of illumination in Polar coordinates (theta and phi) relative to the surface normal must be provided by the user.  The user provides limits and resolution for each of these varialbes in the main function file.  Create_Variable_Combinations.m then takes this information and generates a matrix of all possible positional configurations.  
+The position of the light source is defined in Cartesian coordinates relative to the centre of the illuminated area.  Further to this, a range of allowed angles of illumination in Polar coordinates (theta and phi) relative to the surface normal must be provided by the user.  The user provides limits and resolution for each of these variables in the main function file.  Create_Variable_Combinations.m then takes this information and generates a matrix of all possible positional configurations.  
 
 **Exclude_Unallowed_Combinations.m**
 
@@ -34,16 +34,15 @@ In order to calculate the illumination profile, the far field data for the light
 
 **Illumination_Calculations.m**
 
-Illimination_Calculations.m uses the far field data as well as user provided values to calculate the illumination profile for the light sources in each positional configureation.  In the main function file, the user must provide the name of the varialbe containing the far field data as well as the size of the illuminated area.  Additionally, the user provides the x and y resolution which defines the coarseness of the grid as well as the area over which the flux is calculated.  The position of the camera must also be provided.  This function file assumes that there are four identical light sources positioned across two lines of symmetry.  This can be adjusted as needed.  
+Illimination_Calculations.m uses the far field data as well as user provided values to calculate the illumination profile for the light sources in each positional configuration.  In the main function file, the user must provide the name of the variable containing the far field data as well as the size of the illuminated area.  Additionally, the user provides the x and y resolution which defines the coarseness of the grid as well as the area over which the flux is calculated.  The position of the camera must also be provided.  This function file assumes that there are four identical light sources positioned across two lines of symmetry.  This can be adjusted as needed.  
 
 **Convhull_Option_Reduction.m** 
 
-Convhull_Option_Reduction.m uses a modified convex hull operation to reduce the full matrix of positional configurations to only the most viable by using two figures of merit: the total flux and the standard deviation.  The user can decide to use absolute standard deviation or standard deviation as a percentage of the mean flux.  As standard deviation is not necessarily a perfect descriptor for uniformity, additional data points may be of interest. Therefore, the user can provide tolerance values for both total radiant flux and
-standard deviation to broaden the domain of selected configurations.
+Convhull_Option_Reduction.m uses a modified convex hull operation to reduce the full matrix of positional configurations to only the most viable by using two figures of merit: the total flux and the standard deviation.  The user can decide to use absolute standard deviation or standard deviation as a percentage of the mean flux.  As standard deviation is not necessarily a perfect descriptor for uniformity, additional data points may be of interest. Therefore, the user can provide tolerance values for both total radiant flux and standard deviation to broaden the domain of selected configurations.
 
 **Plot_Save_Results.m**
 
-To visualise the illumination profile of the remaining positional configureations, Plot_Save_Results.m will provide a colormap represening illumination profile and the relative positions of the light sources.  It will also provide the positional information of the light source, the total flux, and the standard deviation of the flux.   
[…]
