### 2019-05-03 3e51b15
Delete Exclude_Unallowed_Combinations.m
