### 2019-05-03 a275ab8
Delete Illumination_Calculations.m
