+++ Illumination_Calculations.m
@@ -1,198 +1,0 @@
-function [Illumination_Data] = Illumination_Calculations(Combinations, Far_Field_Data, Illuminated_Area_Limits, Flux_Data, Camera_Height, Camera_Position)
-
-%Set up structure for data
-Illumination_Data = struct;
-
-%The far field data is loaded
-Far_Field_x = Far_Field_Data(:,1);
-Far_Field_y = Far_Field_Data(:,2);
-
-%These are the user-defined limits for the grid describing the illuminated
-%area
-X_limit = Illuminated_Area_Limits(1);
-X_spacing = Illuminated_Area_Limits(2);
-Y_limit = Illuminated_Area_Limits(3);
-Y_spacing = Illuminated_Area_Limits(4);
-
-%Formation of the grid
-x_grid = -1 * X_limit:X_spacing:X_limit;
-y_grid = -1 * Y_limit:Y_spacing:Y_limit;
-
-%Determines the total number of combinations being investigated
-S = size(Combinations);
-number_combinations = S(1);
-
-%Variables are pre-allocated for speed. 
-data_theta = zeros(number_combinations,1);
-data_phi = zeros(number_combinations,1);
-data_x_position = zeros(number_combinations,1);
-data_y_position = zeros(number_combinations,1);
-data_height = zeros(number_combinations,1);
-
-data_Std_Dev = zeros(number_combinations,1);
-data_Total_Flux = zeros(number_combinations,1);
-data_Std_Perc = zeros(number_combinations, 1);
-
-%The data_Flux_total variable can be very large and require excessive
-%memory.  As it is not necessary for optimisation is can be excluded.  It
-%is not recommended to record this data when investigating large number of
-%combinations.  
-if Flux_Data == 1
-    data_Flux_total = zeros(length(y_grid)-1, length(x_grid)-1, number_combinations);
-end
-
-theta_combinations = Combinations(:,1);
-phi_combinations = Combinations(:,2);
-x_position_combinations = Combinations(:,3);
-y_position_combinations = Combinations(:,4);
-height_combinations = Combinations(:,5);
-
-parfor option = 1:number_combinations
-    
-    %Selects correct value for each variable for each combination
-    theta = theta_combinations(option);
-    phi = phi_combinations(option);
-    x_position = x_position_combinations(option);
-    y_position = y_position_combinations(option);
-    height = height_combinations(option);
-    
-    %Determines the point at which the primary axis of the light source 
-    %intersects the illuminated area
-    d_along_plate = tan(theta)*height;
-    y_along_plate = sin(phi)*d_along_plate;
-    x_along_plate = cos(phi)*d_along_plate;
-
-    light_centre_x = x_position + x_along_plate;
-    light_centre_y = y_position + y_along_plate;
-
-    %Pre-allocates the flux grid 
-    Flux = zeros(length(y_grid)-1, length(x_grid)-1);
-
-    for i = 1:length(x_grid)-1
-        for j = 1:length(y_grid)-1
-
-            %For graphing purposes, the x,y coordinate for each section of
-            %the grid is recorded as its centre
-            x = (x_grid(i) + x_grid(i+1))/2;
-            y = (y_grid(j) + y_grid(j+1))/2;
-
-            %Determines the distance from the source
-            true_distance_from_source = pdist2([x, y, 0], [x_position, y_position, height]);
-            
-            %Determines the reduction in intensity from the source due to
-            %distance
-            Intensity_distance = 1/(true_distance_from_source)^2;
-
-            %Determines the angle from the central beam of light to the
-            %grid position
-            v1 = [x, y, 0] - [x_position, y_position, height];
-            v2 = [light_centre_x, light_centre_y, 0] - [x_position, y_position, height];
-            CosTheta = dot(v1,v2)/(norm(v1)*norm(v2));
[…]
