### 2019-05-03 f6797fc
Delete Optimising_Light_Source_Positioning.m
