### 2019-05-08 #1392 @adrena-lab
Dear @trallard, @cmccomb and @nicoguaro

thank you for handling the manuscript and for the very constructive feedback to our work. We have in the meantime uploaded a revised submission on github. We restructured the organisation of the files in the repository and made major changes to the file readme.md:

1) We expanded the installation instructions to include a description of what each file does.
2) We have added a section on dependencies.
3) We have added/expanded the Example Usage Section to include a description of the example case and how it is carried out. Further, we note that the software is automatically set up to run this example to act as an automated test. Anyone who runs the code as provided will get the same results as shown in the example.
4) We have also added some notes on the community guidelines.
5) We have made a release version v1.0 to package the software

Please note that
1) We received an error message from whedon about the pdf file compilation. Whedon was able to compile the pdf previously and since then, we did not change anything to the relevant file paper.md. Would you be able to resolve this issue?
2) Please note that we have no previous experience in the use of Octave. Based on the comments of @nicoguaro, it seems to be relatively straight forward to run our files, but we feel unable to facilitate this and would prefer to leave the adaptation to other platforms to the user.

We hope to have fully addressed all points raised by the reviewers and look forward to your response.
Thank you and best wishes,
@adrena-lab
