### 2019-05-09 #1392 @nicoguaro
I tested the software using MATLAB 2019a and obtained the results presented in the example. Nevertheless, I obtained the following warning message:

```matlab
Warning: Polynomial is badly conditioned. Add points with distinct X values, reduce the degree of the polynomial, or try centering
and scaling as described in HELP POLYFIT.
> In polyfit (line 79)
  In Convhull_Option_Reduction (line 91)
  In Optimising_Light_Source_Positioning (line 188)
Warning: Polynomial is badly conditioned. Add points with distinct X values, reduce the degree of the polynomial, or try centering
and scaling as described in HELP POLYFIT.
> In polyfit (line 79)
  In Convhull_Option_Reduction (line 91)
  In Optimising_Light_Source_Positioning (line 188)
```

## Comments

- Regarding the *Statement of need* it is present in the first paragraph of the README file, but it might be clearer if it is presented separated.
- The README present details that are too specific. It would be a good idea to move them to documentation files and leave the instructions for installation and execution of the example provided.
- The functions have plenty of comments but it lacks a [help block](https://www.mathworks.com/help/matlab/matlab_prog/add-help-for-your-program.html), especially, defining input and output parameters.
- There are lines of code (and comments) that are too long.
