+++ README.md
@@ -32,7 +32,7 @@ Complete help blocks for each function have been provided which explicitly defin
 
 **Optimising_Light_Source_Positioning.m**
     
-This is the main function file for the software.  In this file, the user provides the required information for each of the following functions. This includes, but is not limited to, the ranges for each positional variable, details regarding the imaging system specifications such as the size of the illuminated area and the position of the camera, and the tolerances for selecting the most viable configurations.  Documentation within this file advises the user where to change the code to adapt it to their specific needs. 
+This is the main script file for the software.  In this file, the user provides the required information for each of the following functions. This includes, but is not limited to, the ranges for each positional variable, details regarding the imaging system specifications such as the size of the illuminated area and the position of the camera, and the tolerances for selecting the most viable configurations.  Documentation within this file advises the user where to change the code to adapt it to their specific needs. 
 
 **Create_Variable_Combinations.m**
 
