+++ Code/Optimising_Light_Source_Positioning.m
@@ -1,7 +1,11 @@
-%This is the main file for the optimisation of light source positioning, it
-%will cover each step of the process along with associated instructions on
-%each step and the modifications the user needs to perform in order to
-%tailor the system to their own set-up 
+%This is the main script file for the optimisation of light source 
+%positioning. In this file, the user provides the required information for 
+%each of the following functions. This includes but is not limited to the
+%ranges for each positional variable, details regarding the imaging system
+%specifications such as the size of the illuminated area and the position
+%of the camera, and the tolerances for selecting the most viable
+%configurations. Documentation within this file advises the user where to
+%change the code to adapt it to their specific needs.
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -17,7 +21,8 @@
 %NOTE: These values are for one light source, in the
 %Illumination_Calculations step (Step 3) it is assumed that there are four
 %identical light sources which are placed around the illuminated area over 
-%two lines of symmetry. It is noted in Step 3 where the user may adjust this.  
+%two lines of symmetry. It is noted in Step 3 where the user may adjust 
+%this.  
 
 
 %The user must adjust these values accordingly:
@@ -49,7 +54,8 @@ H_upper_limit = 110;
 
 %The values provided by the user are collected here into a matrix of input
 %variables which will be used to determine all possible combinations
-Input_Variables = [theta_range_interval, theta_range_lower_limit, theta_range_upper_limit;...
+Input_Variables = [theta_range_interval, theta_range_lower_limit, ...
+    theta_range_upper_limit;...
     phi_range_interval, phi_range_lower_limit, phi_range_upper_limit;...
     x_interval, x_lower_limit, x_upper_limit;...
     y_interval, y_lower_limit, y_upper_limit;...
@@ -70,7 +76,8 @@ disp('Step 1 is Complete')
 %Depending upon the lighting system, some of the combinations will be 
 %unfeasible.  This may be due to design limitations such as providing an 
 %unobstructed view for an imaging device or because certain combinations 
-%are inherently inefficient and removing them will reduce the computing time.  
+%are inherently inefficient and removing them will reduce the computing 
+%time.  
 
 %As this process is highly individual to each lighting system, we
 %have provided some general protocols for the removal of unsuitable
@@ -96,7 +103,7 @@ disp('Step 2 is Complete')
 %angle from the principal axis of the light source (using far field data).
 %The code then utilises a cosine flux correction to account for the 
 %projected area difference between the light path and surface normal.
-%Finally, Lambert’s cosine emission law (and a secondary target-to-lens 
+%Finally, Lambertâ€™s cosine emission law (and a secondary target-to-lens 
 %distance factor) is applied in order to determine the radiant intensity as 
 %observed by the imaging objective.
 
@@ -108,12 +115,12 @@ disp('Step 2 is Complete')
 
 %The far field data is required for the calculation, this will typically be
 %provided by the supplier of the light source (or measured using a
-%goniophotometer). It can be uploaded in any way but should be a two column
-%array where the first column is the angle and the second is the relative 
-%intensity.  We provide the far field data for a 5mm diameter hemispherical
-%LED as an example.
+%goniophotometer). It should be a two column array where the first column
+%is the angle and the second is the relative intensity.  This array should
+%be named Far_Field_Data.  We provide the far field data for a 5mm diameter
+%hemispherical LED as an example.
 
-load Far_Field_Data.mat;
[…]
