+++ Code/Convhull_Option_Reduction.m
@@ -1,7 +1,51 @@
-function [Illumination_Data_Reduced] = Convhull_Option_Reduction(Illumination_Data, Std_Dev_Selector, Total_Flux_Tolerance, Standard_Deviation_Tolerance)
+function [Illumination_Data_Reduced] = Convhull_Option_Reduction(...
+    Illumination_Data, Std_Dev_Selector, Total_Flux_Tolerance, ...
+    Standard_Deviation_Tolerance)
 
-%This function uses a modified convex hull method to select those points which
-%maximize total flux for a given standard deviation. 
+%Overview:  
+%   Convhull_Option_Reduction.m uses a modified convex hull operation to 
+%   reduce the full matrix of positional configurations to only the most 
+%   viable by using two figures of merit: the total flux and the standard 
+%   deviation. Based on this criteria, this function uses a modified convex
+%   hull method to select those points which maximize total flux for a
+%   given standard deviation. 
+%
+%Inputs:
+%   Illumination_Data
+%       This function requires the structure Illumination_Data calculated
+%       in the function Illumination_Calculations.m. It contains the
+%       positional information for each combination as well as the
+%       respective values for total flux, standard deviation, and the
+%       standard deviation as a percentage of the mean flux.  
+%
+%   Std_Dev_Selector
+%       The user can decide to use absolute standard deviation or standard
+%       deviation as a percentage of the mean flux as the second figure of
+%       merit. To choose the absolute standard deviation, the 
+%       Std_Dev_Selector variable should be set to 1, for standard
+%       deviation as a percentage of the mean flux it should be set to 0. 
+%
+%   Total_Flux_Tolerance
+%       As standard deviation is not necessarily a perfect descriptor for
+%       uniformity, additional data points beyond what is selected by the
+%       traditional convex hull function may be of interest. Therefore, the
+%       user can provide tolerance values for both total radiant flux and
+%       standard deviation.  The Total_Flux_Tolerance represents how far
+%       from the convex hull points with regards to total flux a
+%       configuration can be and still be presented as a viable option.  
+%
+%   Standard_Deviation_Tolerance
+%       This variable is akin to the Total_Flux_Tolerance but represents
+%       how far from the convex hull points with regard to standard
+%       deviation a configuration can be and still be presented as a viable
+%       option. 
+%
+%Output: 
+%   Illumination_Data_Reduced
+%       This is a structure containing the positional information for each
+%       combination as well as the respective values for total flux,
+%       standard deviation, and the standard deviation as a percentage of
+%       the mean flux for the reduced options.  
 
 %These lines of code implement the user's selection between the absolute
 %standard deviation and the standard deviation as a percentage of
@@ -65,12 +109,12 @@ Ks = std_dev(K);
 KI = I(K);
 
 %The goal is to achieve a high total intensity with a low standard
-%deviation. Therefore, any points that correspond to a high standard deviation
-%and low total intensity can be disregarded.  As the convhull() function
-%chooses points to encompass the entire range of the data, there will be
-%some points that represent a high standard deviation and low intensity.
-%The following lines of code identify these and removes them from the data
-%set.  
+%deviation. Therefore, any points that correspond to a high standard 
+%deviation and low total intensity can be disregarded.  As the convhull() 
+%function chooses points to encompass the entire range of the data, there 
[…]
