### 2019-05-10 339b1a7
Update Create_Variable_Combinations.m
Update on Help Block
