### 2019-05-10 3beb7bb
Update on Help Block; Minor Bug Fix
