+++ Code/Plot_Save_Results.m
@@ -1,5 +1,45 @@
-function Plot_Save_Results(Illumination_Data_Reduced, Plotted_Area_Limits, Illuminated_Area_Limits, Far_Field_Data, Camera_Position)
-    
+function Plot_Save_Results(Illumination_Data_Reduced,...
+    Plotted_Area_Limits, Illuminated_Area_Limits, Far_Field_Data,...
+    Camera_Position)
+
+%Overview:
+%   To visualise the illumination profile of the remaining positional 
+%   configurations, Plot_Save_Results.m will provide a colormap 
+%   representing illumination profile and the relative positions of the 
+%   light sources. It will also provide the positional information of the
+%   light source, the total flux, and the standard deviation of the flux.
+%
+%Inputs:
+%	Illumination_Data_Reduced
+%       Structure provided by Convhull_Option_Reduction.m.  This contains
+%       the positional information for each configuration as well as the
+%       total flux, standard deviation, and standard deviation as a
+%       percentage of the mean flux.
+%
+%	Plotted_Area_Limits
+%       This input is a vector specified in 
+%       Optimising_Light_Source_Positioning.m.  It contains the x and y
+%       limits of the area of interest for visualisation as well as the
+%       resolution for the calculations.  This region may be larger than
+%       the illuminated area.  
+%
+%	Illuminated_Area_Limits
+%       This input is a vector specified in 
+%       Optimising_Light_Source_Positioning.m for
+%       Illumination_Calculations.m.  These limits describe the region over
+%       which even and intense illumination is desired.  
+%
+%	Far_Field_Data
+%       The far field data is required.  The variable name for the data was 
+%       provided for Illumination_Calculations.m.
+%
+%	Camera_Position
+%       The x,y,z position of the camera/imaging device.
+%
+%Output:
+%   Figure as described in the overview which is saved to the current
+%   directory
+
 X_limit = Plotted_Area_Limits(1);
 X_spacing = Plotted_Area_Limits(2);
 Y_limit = Plotted_Area_Limits(3);
@@ -39,7 +79,8 @@ for option = 1:number_combinations
     y_position = Illumination_Data_Reduced.y_position(option);
     height = Illumination_Data_Reduced.height(option);
     std_data = Illumination_Data_Reduced.standard_deviation(option);
-    std_perc = Illumination_Data_Reduced.standard_deviation_percentage(option);
+    std_perc = ...
+        Illumination_Data_Reduced.standard_deviation_percentage(option);
     flux_data = Illumination_Data_Reduced.total_flux(option);
 
 
@@ -62,7 +103,8 @@ for option = 1:number_combinations
             y = (y_grid(j) + y_grid(j+1))/2;
 
             %Determines the distance from the source
-            true_distance_from_source = pdist2([x, y, 0], [x_position, y_position, height]);
+            true_distance_from_source = pdist2([x, y, 0], [x_position,...
+                y_position, height]);
             
             %Determines the reduction in intensity from the source due to
             %distance
@@ -71,7 +113,8 @@ for option = 1:number_combinations
             %Determines the angle from the central beam of light to the
[…]
