### 2019-05-10 #1392 @adrena-lab
Dear @trallard, @cmccomb and @nicoguaro
thanks for the feedback we received on our recent revision, in particular the comments by @nicoguaro. We have now further revised our contribution as follows:
	•	We fixed a Minor bug in Convhull_Option_Reduction.m to remove warning message
	•	We fixed a minor bug in Plot_Save_Results.m to ensure the rectangle denoting the illuminated area and circles that represent the positions of the light sources are fully visible and not obstructed by the surface plot.
	•	We added a statement of need section to README.md for clarity
	•	We reformatted README.md file to separate installation instructions from the description of each .m & .mat file
	•	We added help block to each function file with information defining the input and output parameters
	•	We edited the code to ensure all lines of code and comments are not too long
	•	Created new release (v1.0.1) to cover addition of help blocks and aforementioned bug fixes
Our repository on Github is updated accordingly.
Again, we are extremely grateful for the time spent on reviewing and testing our JOSS submission and for the constructive feedback that we received during the revision process. We hope that the contribution is now suitable for publication in JOSS.
Best wishes,
Niamh, Alaric and Stefan from @adrenlab
