### 2019-05-10 #1392 @nicoguaro
I would recommend the paper for publication now.
