### 2019-05-13 #1392 @trallard
Hey  @adrena-lab it seems all is now good for publication, can you please go over this list of actions before I move onto the process:

1. Create a new release on GitHub
2. Publish the package in Zenodo and share the DOI here (the general one pointing to the code and all the versions, not the one corresponding to the latest release)
3. Check the paper for: links, authors and DOIs and call `@whedon generate pdf`

Once this is completed please ping me directly
