### 2019-05-15 #1392 @adrena-lab
Dear @trallard,

We created a new release on github (v1.0.1) to encompass the latest revisions. We have published the package on Zenodo and obtained the following DOI:10.5281/zenodo.2847906.  Unfortunately, there appear to be some issues with compiling the .pdf and I am unsure how to proceed.  I would appreaciate any advice.

Thank you,

Niamh
