### 2019-05-15 #1392 @arfon
@adrena-lab - please update the URLs for images in your `paper.md` to link to GitHub RAW URLs e.g. `https://raw.githubusercontent.com/adrena-lab/Optimising-Light-Source-Positioning/Code/Figures/Figure1.png`
