+++ paper.bib
@@ -11,10 +11,10 @@ title = {{Nature of Light}},
 year = {2018}
 }
 
-
-@misc{Convhull,
+@misc{MATLAB&Simulink-MathWorks,
 keywords = {Convhull},
-mendeley-tags = {Convhull},
-title = {{Convhull - MATLAB {\&} Simulink - MathWorks, United Kingdom}},
-howpublished = {https://uk.mathworks.com/help/matlab/ref/convhull.html (accessed Feb 25, 2019)}
+author = {{MATLAB {\&} Simulink - MathWorks}},
+title = {{Convhull}},
+url = {https://uk.mathworks.com/help/matlab/ref/convhull.html},
+urldate = {Feb 25, 2019}
 }
