+++ paper.bib
@@ -12,7 +12,6 @@ year = {2018}
 }
 
 @misc{MATLAB&Simulink-MathWorks,
-keywords = {Convhull},
 author = {{MATLAB {\&} Simulink - MathWorks}},
 title = {{Convhull}},
 url = {https://uk.mathworks.com/help/matlab/ref/convhull.html},
