### 2019-05-16 cd41255
Update paper.md
