+++ paper.bib
@@ -11,7 +11,7 @@ title = {{Nature of Light}},
 year = {2018}
 }
 
-@online{MathWorks,
+@incollection{MathWorks,
 author = {{MATLAB {\&} Simulink - MathWorks}},
 title = {{Convhull}},
 year = {2018},
