+++ paper.bib
@@ -11,9 +11,9 @@ title = {{Nature of Light}},
 year = {2018}
 }
 
-@incollection{MathWorks,
-author = {{MATLAB {\&} Simulink - MathWorks}},
-title = {{Convhull}},
+@misc{MathWorks,
+author = {{MathWorks}},
+title = {{Convhull:  User's Guide (R2018b)}},
 year = {2018},
 url = {https://uk.mathworks.com/help/matlab/ref/convhull.html},
 urldate = {Feb 25, 2019}
