+++ paper.bib
@@ -11,7 +11,7 @@ title = {{Nature of Light}},
 year = {2018}
 }
 
-@misc{MathWorks,
+@online{MathWorks,
 author = {{MathWorks}},
 title = {{Convhull:  User's Guide (R2018b)}},
 year = {2018},
