+++ paper.bib
@@ -13,9 +13,8 @@ year = {2018}
 
 @online{MathWorks,
 author = {{MathWorks}},
-title = {{Convhull:  User's Guide (R2018b)}},
+title = {{Convhull:  User's Guide (R2018b).  Accessed: Feb 25, 2019}},
 year = {2018},
 url = {https://uk.mathworks.com/help/matlab/ref/convhull.html},
 urldate = {Feb 25, 2019}
-note = {Accessed: Feb 25, 2019}
 }
