### 2019-05-16 #1392 @adrena-lab
> Hey @adrena-lab it seems all is now good for publication, can you please go over this list of actions before I move onto the process:
>
> 1. Create a new release on GitHub
> 2. Publish the package in Zenodo and share the DOI here (the general one pointing to the code and all the versions, not the one corresponding to the latest release)
> 3. Check the paper for: links, authors and DOIs and call `@whedon generate pdf`
>
> Once this is completed please ping me directly

Hi @trallard,

we are done with all of the above. Only issue is that the link to the archive is not working. We had shared the details of Zenodo submission in our thread (DOI:10.5281/zenodo.2847906; https://zenodo.org/record/2847906#.XN1OA9NKhdA).

Can this be resolved? If so, I believe we should be ready to publish.

Thanks,
Stefan and Niamh
