### 2019-05-20 #1392 @adrena-lab
Hi @trallard,

I was wondering whether there is an update on our paper as it seems we have addressed all outstanding issues (besides the minor comment that the link to the archive is not working). It would be great to see this work published.

Thank you and kind regards,
Stefan
