### 2019-05-23 #1392 @trallard
@adrena-lab  all looks good to me so I am going to proceed for acceptance 🙋🏻‍♀️

@cmccomb  and @nicoguaro  thanks a lot for your time and valuable contribution to JOSS as reviewers for this submission 🙌🏻

@openjournals/joss-eics this submission has been accepted and is ready to be published 🎉🙌✨
