### 2019-05-23 #1392 @danielskatz
The DOI does not yet resolve for me - once it does, I will close this and the paper will be finalized.
